Fix compiler warning about implict cast to double

+++ tests/auto/corelib/global/qrandomgenerator/tst_qrandomgenerator.cpp
@@ -885,7 +885,7 @@ void tst_QRandomGenerator::stdUniformRealDistribution_data()
     newRow(0, 1);   // canonical
     newRow(0, 200);
     newRow(0, numeric_limits<quint32>::max() + 1.);
-    newRow(0, numeric_limits<quint64>::max() + 1.);
+    newRow(0, double(numeric_limits<quint64>::max()) + 1.);
     newRow(-1, 1.6);
 }
